create truncateDecimals helper for common use in supportTable and cableSpanManip form
