+++ src/app/features/study/presentation/components/sections-tab/newSectionModal/manualSection/supportsTable/supportsTable.component.spec.ts
@@ -428,40 +428,6 @@ describe('SupportsTableComponent', () => {
     });
   });
 
-  describe('truncateDecimals', () => {
-    const makeEvent = (value: string) => ({ target: { value } as HTMLInputElement }) as unknown as Event;
-
-    it('should do nothing when value has no decimal separator', () => {
-      const event = makeEvent('123');
-      component.truncateDecimals(event);
-      expect((event.target as HTMLInputElement).value).toBe('123');
-    });
-
-    it('should do nothing when value has exactly 2 decimal places', () => {
-      const event = makeEvent('1.23');
-      component.truncateDecimals(event);
-      expect((event.target as HTMLInputElement).value).toBe('1.23');
-    });
-
-    it('should do nothing when value has fewer than 2 decimal places', () => {
-      const event = makeEvent('1.2');
-      component.truncateDecimals(event);
-      expect((event.target as HTMLInputElement).value).toBe('1.2');
-    });
-
-    it('should truncate to 2 decimal places when value has more', () => {
-      const event = makeEvent('1.234');
-      component.truncateDecimals(event);
-      expect((event.target as HTMLInputElement).value).toBe('1.23');
-    });
-
-    it('should truncate negative numbers with more than 2 decimal places', () => {
-      const event = makeEvent('-1.234');
-      component.truncateDecimals(event);
-      expect((event.target as HTMLInputElement).value).toBe('-1.23');
-    });
-  });
-
   describe('ngOnInit', () => {
     it('should call getData on init', async () => {
       const getDataSpy = vi.spyOn(component, 'getData');
